Rename shopping list and item components
