+++ src/app/(household)/shoppingLists/[slug]/page.tsx
@@ -1,5 +1,5 @@
 import Card from "@/app/_components/Card";
-import AddItemForm from "@/app/_components/AddItemForm";
+import AddShoppingItemForm from "@/app/_components/AddShoppingItemForm";
 import { api } from "@/trpc/server";
 import ItemsList from "./ItemsList";
 import { ApiName } from "@/server/api/root";
@@ -20,7 +20,7 @@ export default async function Page({ params }: { params: { slug: string } }) {
         </header>
         <div className="mt-5 flex justify-center">
           <Card>
-            <AddItemForm apiName={ApiName.shoppingListItem} listID={shoppingList.listID ?? null} />
+            <AddShoppingItemForm apiName={ApiName.shoppingListItem} listID={shoppingList.listID ?? null} />
           </Card>
         </div>
         <div className="p-4">
