+++ src/app/_components/AddShoppingItemForm.tsx
@@ -10,16 +10,22 @@ import { api } from "@/trpc/react";
 import { useRouter } from "next/navigation";
 import type { ApiName } from "@/server/api/root";
 
+interface iAddItemFormProps {
+  apiName: ApiName.shoppingListItem;
+  // apiName: "shoppingList" 
+  listID: string | null;
+}
 
-export default function AddItemForm({ apiName }: { apiName: ApiName.shoppingList }) {
+
+export default function AddShoppingItemForm(params: iAddItemFormProps) {
   /** As the api mutation in this Component is interacting with a Page (Server Component)
    * we need to use `router.refresh()` to invaliadate the cached data in the Page (Server Component).
    * If the cache data is in a "use client" component then we can use `api.useUtils()` hook instead.
    * further reading here: https://trpc.io/docs/client/react/useUtils
    */
   const router = useRouter();
+  const { apiName, listID } = params;
 
-  // Using the apiName to determine which api to use dynamically.
   const { mutate } = api[apiName].create.useMutation({
     onSuccess: (response) => {
       console.log(`CREATED ${apiName}`, response);
@@ -28,22 +34,27 @@ export default function AddItemForm({ apiName }: { apiName: ApiName.shoppingList
   });
 
   // schema definition can be kept here as its the same for shopping list and chores list
+  // we could pass in more specific schema as a props if needed
   const formSchema = z.object({
-    title: z.string().min(1, { message: "Name is required" }),
+    name: z.string().min(1, { message: "Name is required" }),
   });
 
   // uses the schema and mutate funnction setup above
   function handleSubmit(data: z.infer<typeof formSchema>) {
+    if (listID === null) {
+      console.error("List ID is null, cannot add item");
+      return;
+    }
     console.log("SUBMITTING ITEM TO ADD: ", data);
-    mutate(data);
+    mutate({ ...data, listID });
   }
 
   return (
     <GenericForm formSchema={formSchema} handleSubmit={handleSubmit}>
       <FormItem>
-        <FormLabel fieldName={"title"}>Create a new list</FormLabel>
+        <FormLabel fieldName={"title"}>Create a new item</FormLabel>
         <Inputfield
-          fieldName={"title"}
+          fieldName="name"
           placeholder="Enter a name..."
           className="w-40"
         />
@@ -56,4 +67,3 @@ export default function AddItemForm({ apiName }: { apiName: ApiName.shoppingList
     </GenericForm>
   );
 }
-
