+++ src/app/_components/AddShoppingListForm.tsx
@@ -10,21 +10,16 @@ import { api } from "@/trpc/react";
 import { useRouter } from "next/navigation";
 import type { ApiName } from "@/server/api/root";
 
-interface iAddItemFormProps {
-  apiName: ApiName.shoppingListItem;
-  listID: string | null;
-}
 
-
-export default function AddItemForm(params: iAddItemFormProps) {
+export default function AddShoppingListForm({ apiName }: { apiName: ApiName.shoppingList }) {
   /** As the api mutation in this Component is interacting with a Page (Server Component)
    * we need to use `router.refresh()` to invaliadate the cached data in the Page (Server Component).
    * If the cache data is in a "use client" component then we can use `api.useUtils()` hook instead.
    * further reading here: https://trpc.io/docs/client/react/useUtils
    */
   const router = useRouter();
-  const { apiName, listID } = params;
 
+  // Using the apiName to determine which api to use dynamically.
   const { mutate } = api[apiName].create.useMutation({
     onSuccess: (response) => {
       console.log(`CREATED ${apiName}`, response);
@@ -33,19 +28,14 @@ export default function AddItemForm(params: iAddItemFormProps) {
   });
 
   // schema definition can be kept here as its the same for shopping list and chores list
-  // we could pass in more specific schema as a props if needed
   const formSchema = z.object({
-    name: z.string().min(1, { message: "Name is required" }),
+    title: z.string().min(1, { message: "Name is required" }),
   });
 
   // uses the schema and mutate funnction setup above
   function handleSubmit(data: z.infer<typeof formSchema>) {
-    if (listID === null) {
-      console.error("List ID is null, cannot add item");
-      return;
-    }
     console.log("SUBMITTING ITEM TO ADD: ", data);
-    mutate({ ...data, listID });
+    mutate(data);
   }
 
   return (
@@ -53,7 +43,7 @@ export default function AddItemForm(params: iAddItemFormProps) {
       <FormItem>
         <FormLabel fieldName={"title"}>Create a new list</FormLabel>
         <Inputfield
-          fieldName="name"
+          fieldName={"title"}
           placeholder="Enter a name..."
           className="w-40"
         />
@@ -66,3 +56,4 @@ export default function AddItemForm(params: iAddItemFormProps) {
     </GenericForm>
   );
 }
+
